requirements.txt
torch>=1.13.0
torchvision>=0.14.0
timm>=0.6.12
ultralytics>=8.0.0
opencv-python>=4.6.0
streamlit>=1.18.0
pyttsx3>=2.90
gTTS>=2.3.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.1.0
matplotlib>=3.6.0
seaborn>=0.12.0
albumentations>=1.3.0
tqdm
Pillow
tensorboard
pyyaml
playsound
